Code/requirements.txt
aiohttp==3.8.1
aiohttp-cors==0.7.0
aiorwlock==1.3.0
aiosignal==1.2.0
anaconda==0.0.1.1
ansicon==1.89.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
astroid==2.11.3
asttokens==2.0.5
async-generator==1.10
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.4.0
autopep8==1.6.0
auxlib==0.0.43
backcall==0.2.0
beautifulsoup4==4.10.0
black==22.6.0
bleach==5.0.1
blessed==1.19.1
bokeh==2.4.3
boto3==1.24.43
botocore==1.27.43
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.0.12
cliar==1.3.5
click==8.0.4
click-default-group==1.2.2
cloudpickle==2.1.0
colorama==0.4.4
colorful==0.5.4
commonmark==0.9.1
compound-word-splitter==0.4
conda==4.2.7
cycler==0.11.0
dask==2022.1.1
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
distributed==2022.1.1
Django==4.0.3
django-emoji==2.2.2
dm-tree==0.1.7
emac==0.1.0
emoji==1.7.0
entrypoints==0.4
executing==0.8.3
export==0.2.0
fastapi==0.79.0
fastjsonschema==2.16.1
filelock==3.6.0
flake8==4.0.1
Flask==2.1.3
flask-render-specific-template==1.1
fonttools==4.31.2
frictionless==4.40.7
frozenlist==1.3.0
fsspec==2022.7.1
gast==0.2.2
gh==0.0.4
gitdb==4.0.9
GitPython==3.1.27
google-api-core==2.8.2
google-auth==2.9.1
googleapis-common-protos==1.56.4
gpustat==1.0.0rc1
greenlet==1.1.2
grpcio==1.43.0
gym==0.21.0
h11==0.13.0
HeapDict==1.0.1
html5lib==1.1
idna==3.3
image==1.5.33
imageio==2.21.0
importlib-metadata==4.11.3
iniconfig==1.1.1
ipykernel==6.15.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
isodate==0.6.1
itsdangerous==2.0.1
jason==0.1.7
jedi==0.18.1
Jinja2==3.1.2
jinxed==1.2.0
jmespath==1.0.1
joblib==1.1.0
jsonify==0.5
jsonschema==4.9.0
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
kiwisolver==1.4.2
lazy-object-proxy==1.7.1
live==0.0.3
locket==1.0.0
lxml==4.8.0
lz4==4.0.2
marko==1.2.1
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
matplotlib-venn==0.11.7
mccabe==0.6.1
memory-profiler==0.60.0
mistune==0.8.4
modin==0.15.2
modin-spreadsheet==0.1.2
msgpack==1.0.4
multidict==6.0.2
mypy-extensions==0.4.3
nana==1.0.6
nano==0.10.0
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.5
notebook==6.4.12
numpy==1.22.3
nvidia-ml-py==11.495.46
opencensus==0.10.0
opencensus-context==0.1.2
outcome==1.2.0
packaging==21.3
pandas==1.4.3
pandas-schema==0.3.6
pandera==0.11.0
pandocfilters==1.5.0
parso==0.8.3
partd==1.2.0
pathspec==0.9.0
pbr==0.11.1
petl==1.7.11
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.1
pluggy==1.0.0
plumbum==1.7.2
poetry-semver==0.1.0
prettytable==3.3.0
prometheus-client==0.13.1
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.0
pure-eval==0.2.2
py==1.11.0
py-spy==0.3.12
py4j==0.10.9.5
pyaml==21.10.1
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.1
pyflakes==2.4.0
pygame==2.1.2
Pygments==2.12.0
pyparsing==3.0.7
pyrsistent==0.18.1
PySocks==1.7.1
pyspark==3.3.0
pytest==7.1.1
python-dateutil==2.8.2
python-json-logger==2.0.4
python-slugify==6.1.2
pytz==2022.1
PyWavelets==1.3.0
pywin32==304
pywinpty==2.0.6
PyYAML==6.0
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
ray==1.12.1
ray-cpp==1.12.1
redis==3.5.3
regex==2022.7.25
requests==2.27.1
requirements-detector==1.0.3
rfc3986==2.0.0
rich==12.5.1
rpyc==4.1.5
rsa==4.9
s3transfer==0.6.0
scikit-image==0.19.3
scikit-learn==1.1.1
scipy==1.8.0
scoop==0.7.1.1
seaborn==0.11.2
selenium==4.4.3
Send2Trash==1.8.0
shellingham==1.5.0
simpleeval==0.9.12
six==1.16.0
sklearn==0.0
smart-open==6.0.0
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
splinter==0.17.0
splitter==0.1.1
sqlparse==0.4.2
stack-data==0.3.0
starlette==0.19.1
stringcase==1.2.0
tableschema-to-template==0.0.12
tabulate==0.8.10
tblib==1.7.0
tensorboardX==2.5.1
terminado==0.15.0
tex==1.8
text-unidecode==1.3
threadpoolctl==3.1.0
tifffile==2022.7.31
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.1
traitlets==5.3.0
trio==0.21.0
trio-websocket==0.9.2
typer==0.6.1
typing-inspect==0.8.0
typing_extensions==4.2.0
tzdata==2022.1
unicode==2.8
unicoder==1.1.1017
urllib2-file==0.2.1
urllib3==1.26.9
urwid==2.1.2
uvicorn==0.16.0
validators==0.20.0
valuedispatch==0.0.1
virtualenv==20.14.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
Werkzeug==2.1.2
wget==3.2
widgetsnbextension==3.6.1
wordcloud @ file:///C:/Users/nunto/Downloads/wordcloud-1.8.1-cp39-cp39-win_amd64.whl
wrapt==1.14.0
wsproto==1.1.0
XlsxWriter==3.0.3
yarl==1.8.0
zict==2.2.0
zipp==3.8.0
